fix pir api 2
