+++ test/legacy_test/test_bicubic_interp_v2_op.py
